+++ crates/cli/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Major
 
-- Rename `--output` to `--output-dir` for `rust-applet-{build,test}`
+- Rename `--output` to `--output-dir` for `rust-applet-build`
 - Rename `--serial` to `--protocol` with more support
 - Move `--serial` and `--timeout` to commands that need them