+++ crates/cli/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.2.0-git
+## 0.2.0
 
 ### Major
 
@@ -50,4 +50,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 12 -->
+<!-- Increment to skip CHANGELOG.md test: 0 -->